Move manta/ dir to storage/
