+++ storage/authentication/private_key_signer.go
