+++ storage/authentication/signer.go
