+++ storage/authentication/util.go
