+++ storage/examples/getobject/main.go
