+++ storage/examples/putobject/foo.txt
