+++ storage/examples/putobject/main.go
