+++ storage/signing.go
@@ -1,4 +1,4 @@
-package manta
+package storage
 
 import (
 	"bytes"
